each user can see his own transactions

+++ lib/utils/mongo/transaction.ts
@@ -5,15 +5,16 @@ import { decrementCount, incrementCount, updateSalary } from "./meta";
 
 mongoose.connect(process.env.MONGO_URI!).then(() => console.log("connected"));
 
-export async function getTransactions(start: number) {
+export async function getTransactions(start: number, user: string) {
   start = start === undefined ? 0 : start;
-  return Transaction.find({}).skip(start).limit(5);
+  console.log(user);
+  return Transaction.find({ owner: user }).skip(start).limit(5);
 }
-
 export async function addTransaction(t: transaction) {
   // increment the number of transactions
 
   const newTransaction = new Transaction({
+    owner: t.owner,
     title: t.title,
     type: t.type,
     date: t.date,
@@ -23,19 +24,20 @@ export async function addTransaction(t: transaction) {
 
   return Promise.all([
     newTransaction.save(),
-    updateSalary(t.type == "Expense", t.amount as number),
-    incrementCount(),
+    updateSalary(t.type == "Expense", t.amount as number, t.owner),
+    incrementCount(t.owner),
   ]);
 }
 
 export async function deleteTransaction(
   id: string,
   expense: boolean,
-  amount: number
+  amount: number,
+  user: string
 ) {
   await Promise.all([
-    decrementCount(),
-    Transaction.deleteOne({ _id: id }),
-    updateSalary(!expense, amount),
+    decrementCount(user),
+    Transaction.deleteOne({ _id: id, owner: user }),
+    updateSalary(!expense, amount, user),
   ]);
 }
